test(e2e): move scenario logic out of test wrapper
